requirements.txt
# Core dependencies
haystack-ai>=2.0.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
chromadb>=0.4.22
chroma-haystack>=0.1.2

# Document processing
pypdf>=3.17.1
markdown>=3.5.1
beautifulsoup4>=4.12.0

# Web interface
streamlit>=1.31.0

# Testing
pytest>=7.4.0

# OpenAI integration
openai>=1.12.0

# Optional dependencies
pytest-xdist>=3.5.0
pytest-cov>=4.1.0 